add a example for select
seems if all case condition is not met, then default condition is execed. And during default is execing, if other case condition is met, then it will be blocked. And after default is finished, a new for iteration will start over. And the process described above will go over again.